+++ CHANGELOG.md
@@ -9,6 +9,7 @@ starting after version 4.0.12, but historic entries might not.
 
 ### Added
 
+- Windows: Dynamically monitor ZCM1 and ZCM2 controller connections and disconnections
 - macOS: Add build support for M1/Apple Silicon (Fixes #440)
 - Linux/Debian: Force building tracker and test programs (Fixes #437)
 - Added `psmove_tracker_opencv.h` for OpenCV-specific functions
@@ -24,6 +25,7 @@ starting after version 4.0.12, but historic entries might not.
 
 ### Changed
 
+- Windows: `psmove pair -d` now waits for controller notifications instead of repeatedly rescanning
 - CI: Migrate from Travis CI (macOS, Linux), and AppVeyor (Windows) to Github Actions
 - Linux: Build OpenCV from source
 - `CMakeLists.txt`: Add check for 'git submodule init' (Fixes #352)